sys, prog: switch values to to uint64
We currently use uintptr for all values.
This won't work for 32-bit archs.
Moreover in some cases we use uintptr but assume
that it is always 64-bits (e.g. in encodingexec).
Switch everything to uint64.

Update #324

+++ prog/prog_test.go
@@ -96,7 +96,7 @@ func TestVmaType(t *testing.T) {
 		if len(c.Args) != 6 {
 			t.Fatalf("generated wrong number of args %v", len(c.Args))
 		}
-		check := func(v, l Arg, min, max uintptr) {
+		check := func(v, l Arg, min, max uint64) {
 			va, ok := v.(*PointerArg)
 			if !ok {
 				t.Fatalf("vma has bad type: %v", v)
